src/generic-puller/requirements.txt
coralogix_logger==2.0.5